requirements.txt
pandas
ipdb
requests
tabula
pdfplumber
camelot-py[cv]
PyPDF2==2.12.1
pymupdf
pdf2image
pytesseract